[Menu] Store paging

+++ src/main/java/com/harmoni/pos/business/service/store/StoreService.java
@@ -9,9 +9,9 @@ import java.util.Map;
 public interface StoreService {
 
     int create(StoreDto storeDto);
-    int delete(Long id);
-    Store get(Long id);
-    Map<String, Object> getAllStoresByChainIdPaginated(Long chainId, int page, int size, String search);
-    List<Store> getAllStoresByChainId(Long chainId, String search);
-    int update(Long id, StoreDto storeDto);
+    int delete(Integer id);
+    Store get(Integer id);
+    Map<String, Object> getAllStoresByChainIdPaginated(Integer chainId, int page, int size, String search);
+    List<Store> getAllStoresByChainId(Integer chainId, String search);
+    int update(Integer id, StoreDto storeDto);
 }
